Implementing (Extremely Naive) Wrong Guess Detection.
Basically, if our trial intersections yield 0 possible characters, we
assume the last character we guessed was bad, and backtrack.

This is actually a rather solid solution, but it could be much better.
We could sample the times for each dot outputted by level06 and run
some analysis there (checking to see if the character we're guessing is
where the fork is (based on the time), for example). But this would
either require major refactoring, or some really bad code flow.

As it stands now, timing.py can easily handle passwords up to ~60
characters, at which point things have a very high likelihood of going
pear shaped.